+++ app/src/main/java/com/example/se_2019/Activity/ForumActivity.java
@@ -1,4 +1,4 @@
-package com.example.se_2019;
+package com.example.se_2019.Activity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -10,7 +10,11 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.se_2019.Post;
+import com.example.se_2019.R;
+
 import java.util.ArrayList;
+import java.util.Date;
 
 public class ForumActivity extends AppCompatActivity {
 
@@ -28,15 +32,16 @@ public class ForumActivity extends AppCompatActivity {
         list_itemArrayList = new ArrayList<Post>();
         //여기 목록에 띄워질 애들
         //함수 만들어서 다시 불러와야함
-//        list_itemArrayList.add(
-//                new Post(R.mipmap.ic_launcher,"조현아","제목",new Date(System.currentTimeMillis()),"안녕하세요"));
-//        myListAdapter = new MyListAdapter(ForumActivity.this,list_itemArrayList);
-//        listView.setAdapter(myListAdapter);
+        list_itemArrayList.add(
+                new Post(R.mipmap.ic_launcher,"조현아","제목",new Date(System.currentTimeMillis()),"안녕하세요"));
+        myListAdapter = new MyListAdapter(ForumActivity.this,list_itemArrayList);
+        listView.setAdapter(myListAdapter);
     }
     public void writeClick(View view){
         Intent intent = new Intent(ForumActivity.this, AddPostActivity.class);
         startActivity(intent);
     }
+
 //이 밑은 메뉴
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
@@ -44,7 +49,6 @@ public class ForumActivity extends AppCompatActivity {
         getMenuInflater().inflate(R.menu.menu_main, menu);//여기서 menu/menu_main UI를 가져옴
         return true;
     }
-
 
     @Override
     public boolean onOptionsItemSelected (MenuItem item){
