+++ app/src/main/java/com/example/se_2019/Activity/MyListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.se_2019;
+package com.example.se_2019.Activity;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,6 +7,9 @@ import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.example.se_2019.Post;
+import com.example.se_2019.R;
 
 import java.util.ArrayList;
