+++ app/src/main/java/com/example/se_2019/Activity/RegisterActivity.java
@@ -1,4 +1,4 @@
-package com.example.se_2019;
+package com.example.se_2019.Activity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -13,6 +13,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import com.android.volley.RequestQueue;
 import com.android.volley.Response;
 import com.android.volley.toolbox.Volley;
+import com.example.se_2019.R;
+import com.example.se_2019.DBRequest.RegisterRequest;
 
 import org.json.JSONException;
 import org.json.JSONObject;
