+++ app/src/main/java/com/example/se_2019/DBRequest/AddRoomRequest.java
@@ -1,4 +1,4 @@
-package com.example.se_2019;
+package com.example.se_2019.DBRequest;
 
 import com.android.volley.AuthFailureError;
 import com.android.volley.Response;
@@ -7,11 +7,13 @@ import com.android.volley.toolbox.StringRequest;
 import java.util.HashMap;
 import java.util.Map;
 
-public class RoomRequest extends StringRequest {
-    final static private String URL="http://deu04202.dothome.co.kr/Room.php";
+public class AddRoomRequest extends StringRequest {
+    //방을 추가시킬 때 해당 userID
+    final static private String URL="http://deu04202.dothome.co.kr/RoomAdd.php";
     private Map<String, String> map;
 
-    public RoomRequest(String roomID, String userID, String roomName, String subName, Response.Listener<String> listener){
+
+    public AddRoomRequest(String roomID, String userID, String roomName, String subName, Response.Listener<String> listener){
         super(Method.POST, URL, listener, null);
 
         map = new HashMap<>();
