+++ app/src/main/java/com/example/se_2019/DBRequest/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.example.se_2019;
+package com.example.se_2019.DBRequest;
 
 import com.android.volley.AuthFailureError;
 import com.android.volley.Response;
