+++ app/src/main/res/layout/activity_add_room.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddRoomActivity"
+    tools:context=".Activity.AddRoomActivity"
     >
 
     <com.google.android.material.appbar.AppBarLayout
