+++ app/src/main/res/layout/content_add_post.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:id="@+id/activity_write"
-    tools:context = ".AddPostActivity">
+    tools:context = ".Activity.AddPostActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
