+++ app/src/main/res/layout/content_add_room.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".AddRoomActivity"
+    tools:context=".Activity.AddRoomActivity"
     tools:showIn="@layout/activity_add_room"
     android:orientation="vertical"
     >
