+++ app/src/main/res/layout/content_forum.xml
@@ -4,7 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ForumActivity"
+    app:layout_behavior="@string/appbar_scrolling_view_behavior"
+    tools:context=".Activity.ForumActivity"
+    tools:showIn="@layout/activity_forum"
     android:orientation="vertical">
